refactor: change directory structure in terms of DDD
