+++ src/main/java/nextstep/subway/map/ui/MapController.java
@@ -1,9 +1,12 @@
-package nextstep.subway.map;
+package nextstep.subway.map.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import nextstep.subway.map.dto.MapResponse;
+import nextstep.subway.map.application.MapService;
 
 @RestController
 @RequestMapping("/maps")
